mds: Interval tree implementation
Interval tree is an optimized data structure for representing and querying intervals. Elementary intervals are represented as nodes of an avl tree and the corresponding data is stored on these nodes based on a concept of span. This representation allows log(n) (where n is the number of data) storage. The balanced avl tree allows a log(n) query. The implementation is a template class that is instantiated based on parameters : - Interval type - Data type

Signed-off-by: Jojy George Varghese <jvarghese@scalecomputing.com>